dial: the proxy confines its device connections; tcp.rs is a plain socket wrapper
On FreeBSD the route check now runs inside connect, after socket() and
TCP_NODELAY, so a refused destination costs one socket open.
