+++ src/reflector/dial.rs
@@ -5,7 +5,7 @@
 //! - [`connection`]: the per-connection bidirectional HTTP byte splice: framing, authority rewriting,
 //!   independent per-direction half-close, drop-and-close backpressure.
 //! - [`proxy`]: the per-device reactor [`Handler`](crate::reactor::Handler) that accepts clients, opens
-//!   egress-pinned device connections, and owns a pool of them.
+//!   device connections confined to the target interface ([`egress`]), and owns a pool of them.
 //! - [`rewrite`]: the SSDP-side entry. [`rewrite_location`] rewrites a DIAL discovery message's
 //!   `LOCATION` to a source-side proxy, minting and registering one on demand.
 //!
@@ -14,6 +14,7 @@
 //! sees the advertisements that refresh it.
 
 mod connection;
+mod egress;
 mod proxy;
 mod rewrite;
